Reorganize benchmarks and common files, and add some logic to make config and utils available.
