+++ README.md
@@ -1,3 +1,6 @@
 # About
 
 Benchmarks of cuDF.
+
+Benchmarks of public APIs are contained in the API subdirectory.
+Benchmarks of non-public internals are contained in the internal subdirectory.
